backend/requirements.txt
# Django 核心框架
Django==5.2.2

# API 框架
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7

# 測試框架
pytest==8.2.2
pytest-cov==6.1.1
pytest-django==4.11.1

# CORS 支持
django-cors-headers==4.4.0

# WebSocket 和實時通信
channels[daphne]>=4.2,<4.3
channels-redis<=4.3

# 數據庫適配器
psycopg2-binary==2.9.10

# 安全和生產環境依賴（可選）
# gunicorn==21.2.0  # WSGI 服務器
# whitenoise==6.6.0  # 靜態文件服務